Get Location Data V3

+++ src/main/java/com/eden/eden_crm_sec_crm_back/dto/request/LocationDto.java
@@ -2,10 +2,12 @@ package com.eden.eden_crm_sec_crm_back.dto.request;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class LocationDto {
     String locationName;
     String accessType;
-    Double longitude;
-    Double latitude;
+    BigDecimal longitude;
+    BigDecimal latitude;
 }
